% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andsport-package.r
\docType{package}
\name{andsport}
\alias{andsport}
\alias{andsport-package}
\title{andsport.}
\description{
Data about towns, sport facilities and sport organizations in Andalusia.
}
\details{
To use the datasets:

Andalusian towns -> data(pueblos)
Andalusian sport facilities -> data(instalaciones)
Andalusian sport organizations -> data(entidades)
}
