\name{instalaciones}
\alias{instalaciones}
\docType{data}
\title{Andalusian sport facilities}
\description{Andalusian sport facilities dataset}
\usage{data("instalaciones")}
\format{A data frame.}
\source{https://ws072.juntadeandalucia.es/pdida/buscador.htm}
\examples{
data(instalaciones)
## maybe str(instalaciones) ; plot(instalaciones) ...
}
\keyword{datasets}
